backend/requirements.txt
Flask==3.0.3
flask-cors==5.0.0
faiss-cpu==1.9.0
sentence-transformers==3.1.1
numpy==1.26.4
